// src/app/rest/model/dtosAccountCredentialsDto.ts
/**
 * Operator Automation Backend API
 * Operator Automation Backend API overview.
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


export interface DtosAccountCredentialsDto { 
    password?: string;
    username?: string;
}